backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
openpyxl>=3.1,<4.0
requests>=2.20,<3.0
reportlab>=4.2,<5.0
